chore: add circom tests
