+++ circuits/circom/test/circuits/12_point_sha_256_test.circom
@@ -1,5 +1,5 @@
 pragma circom 2.1.2;
 
-include "../verify_nullifier.circom";
+include "../../verify_nullifier.circom";
 
 component main = sha256_12_coordinates(64, 4);
