+++ circuits/circom/test/circuits/a_div_b_pow_c_test.circom
@@ -1,5 +1,5 @@
 pragma circom 2.1.2;
 
-include "../verify_nullifier.circom";
+include "../../verify_nullifier.circom";
 
 component main = a_div_b_pow_c(64, 4);
