+++ circuits/circom/test/circuits/compression_verification_test.circom
@@ -1,5 +1,5 @@
 pragma circom 2.1.2;
 
-include "../verify_nullifier.circom";
+include "../../verify_nullifier.circom";
 
 component main = verify_ec_compression(64, 4);
