+++ circuits/circom/test/circuits/v1_test.circom
@@ -1,5 +1,5 @@
 pragma circom 2.1.2;
 
-include "../verify_nullifier.circom";
+include "../../verify_nullifier.circom";
 
 component main = plume_v1(64, 4, 29);
