--- circuits/circom/test/hash_to_curve_test.circom
@@ -1,5 +1,0 @@
-pragma circom 2.1.2;
-
-include "../node_modules/secp256k1_hash_to_curve_circom/circom/hash_to_curve.circom";
-
-component main = HashToCurve(62);
